+++ core/src/test/java/org/trade/core/persistent/RoleServiceIT.java
@@ -14,8 +14,14 @@ import org.trade.core.ApplicationProfileInitializer;
 import org.trade.core.ApplicationRepositoryConfig;
 import org.trade.core.TradestrategyBase;
 import org.trade.core.persistent.role.Role;
+import org.trade.core.persistent.role.RoleDTO;
 import org.trade.core.persistent.role.RoleService;
+import org.trade.core.util.JSONMapper;
 
+import java.util.List;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 
@@ -56,7 +62,11 @@ public class RoleServiceIT {
     public void tearDown() throws ClassNotFoundException {
 
         Role role = roleService.findRoleByName(roleName);
-        roleService.deleteRole(role);
+
+        if (null != role) {
+
+            roleService.deleteRole(role);
+        }
     }
 
     /**
@@ -74,5 +84,36 @@ public class RoleServiceIT {
         role = new Role(roleName, roleName);
         role = roleService.saveRole(role);
         assertNotNull(role.getId());
+    }
+
+    @Test
+    public void findAllTopLevelRoles() {
+
+        List<RoleDTO> roles = roleService.findAllTopLevelRoleDTOs();
+        assertFalse(roles.isEmpty());
+    }
+
+    @Test
+    public void findRoleByName() {
+
+        RoleDTO role = roleService.findRoleDTOByName(Role.ROLE_ADMIN);
+        assertNotNull(role);
+
+        // Manager role
+        assertFalse(role.getContainRoleDTOs().isEmpty());
+        assertEquals(Role.ROLE_MANAGER, role.getContainRoleDTOs().getFirst().getName());
+
+        // User role
+        assertFalse(role.getContainRoleDTOs().getFirst().getContainRoleDTOs().isEmpty());
+        assertEquals(Role.ROLE_USER, role.getContainRoleDTOs().getFirst().getContainRoleDTOs().getFirst().getName());
+    }
+
+    @Test
+    public void findRoleByNameDTO() {
+
+        Role role = roleService.findRoleByName(Role.ROLE_MANAGER);
+        assertNotNull(role);
+        RoleDTO roleDTO = JSONMapper.convertEntityToDTO(role, RoleDTO.class);
+        assertNotNull(roleDTO);
     }
 }
